PWA(Install Web App): had to move all files in service out to be able to download App
